pyproject.toml

[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.49.0",
    "autogen-agentchat>=0.4.9.2",
    "autogen-ext[grpc,mcp,ollama,openai]>=0.4.9.2",
    "bs4>=0.0.2",
    "gradio>=5.22.0",
    "httpx>=0.28.1",
    "ipywidgets>=8.1.5",
    "langchain>=1.2.10",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-core>=1.2.12",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=0.3.9",
    "langchain-qdrant>=1.1.0",
    "langfuse>=3.14.1",
    "langgraph>=0.3.18",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langsmith>=0.3.18",
    "lxml>=5.3.1",
    "mcp-server-fetch>=2025.1.17",
    "mcp[cli]>=1.5.0",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "playwright>=1.51.0",
    "plotly>=6.0.1",
    "polygon-api-client>=1.14.5",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.11",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.3",
    "semantic-kernel>=1.25.0",
    "sendgrid>=6.11.0",
    "sentence-transformers>=5.2.2",
    "setuptools>=78.1.0",
    "smithery>=0.1.0",
    "speedtest-cli>=2.1.3",
    "streamlit>=1.54.0",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.2.4",
    "mcp-use>=1.6.0",
    "ddgs>=9.10.0",
    "agno>=2.5.2",
    "run>=0.3",
    "langchain-agent>=0.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "google-adk[eval]>=1.25.1",
    "crewai>=1.6.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
